requirements.txt
Django==1.6.6
South==1.0
certifi==14.05.14
django-bower==5.0.1
django-classy-tags==0.5.1
django-cms==3.0.5
django-mptt==0.6.1
django-sekizai==0.7
djangocms-admin-style==0.2.2
gunicorn==19.1.1
html5lib==0.999
requests==2.4.0
six==1.7.3
